feat: rename payment
