+++ app/types/module.py
@@ -4,10 +4,9 @@ from uuid import UUID
 from fastapi import APIRouter
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from app.core.checkout import schemas_checkout
 from app.core.groups.groups_type import AccountType, GroupType
-from app.core.mypayment import schemas_mypayment
 from app.core.notification.schemas_notification import Topic
-from app.core.payment import schemas_payment
 from app.core.permissions.type_permissions import ModulePermissions
 from app.types.factory import Factory
 
@@ -20,7 +19,7 @@ class CoreModule:
         factory: Factory | None,
         router: APIRouter | None = None,
         payment_callback: Callable[
-            [schemas_payment.CheckoutPayment, AsyncSession],
+            [schemas_checkout.CheckoutPayment, AsyncSession],
             Awaitable[None],
         ]
         | None = None,
@@ -46,7 +45,7 @@ class CoreModule:
         self.root = root
         self.router = router or APIRouter(tags=[tag])
         self.payment_callback: (
-            Callable[[schemas_payment.CheckoutPayment, AsyncSession], Awaitable[None]]
+            Callable[[schemas_checkout.CheckoutPayment, AsyncSession], Awaitable[None]]
             | None
         ) = payment_callback
         self.mypayment_callback: (
@@ -67,7 +66,7 @@ class Module(CoreModule):
         default_allowed_account_types: list[AccountType] | None = None,
         router: APIRouter | None = None,
         payment_callback: Callable[
-            [schemas_payment.CheckoutPayment, AsyncSession],
+            [schemas_checkout.CheckoutPayment, AsyncSession],
             Awaitable[None],
         ]
         | None = None,
@@ -97,7 +96,7 @@ class Module(CoreModule):
         self.default_allowed_account_types = default_allowed_account_types
         self.router = router or APIRouter(tags=[tag])
         self.payment_callback: (
-            Callable[[schemas_payment.CheckoutPayment, AsyncSession], Awaitable[None]]
+            Callable[[schemas_checkout.CheckoutPayment, AsyncSession], Awaitable[None]]
             | None
         ) = payment_callback
         self.mypayment_callback: (
