+++ tests/core/test_checkout.py
@@ -14,9 +14,9 @@ from helloasso_python.models.hello_asso_api_v5_models_carts_init_checkout_respon
 from pytest_mock import MockerFixture
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.payment import cruds_payment, models_payment, schemas_payment
-from app.core.payment.payment_tool import PaymentTool
-from app.core.payment.types_payment import HelloAssoConfig, HelloAssoConfigName
+from app.core.checkout import cruds_checkout, models_checkout, schemas_checkout
+from app.core.checkout.payment_tool import PaymentTool
+from app.core.checkout.types_checkout import HelloAssoConfig, HelloAssoConfigName
 from app.core.schools import schemas_schools
 from app.core.users import schemas_users
 from app.types.module import Module
@@ -30,9 +30,9 @@ from tests.commons import (
 if TYPE_CHECKING:
     from app.core.utils.config import Settings
 
-checkout_with_existing_checkout_payment: models_payment.Checkout
-existing_checkout_payment: models_payment.CheckoutPayment
-checkout: models_payment.Checkout
+checkout_with_existing_checkout_payment: models_checkout.Checkout
+existing_checkout_payment: models_checkout.CheckoutPayment
+checkout: models_checkout.Checkout
 
 user_schema: schemas_users.CoreUser
 
@@ -43,7 +43,7 @@ TEST_MODULE_ROOT = "tests"
 async def init_objects() -> None:
     global checkout_with_existing_checkout_payment
     checkout_with_existing_checkout_payment_id = uuid.uuid4()
-    checkout_with_existing_checkout_payment = models_payment.Checkout(
+    checkout_with_existing_checkout_payment = models_checkout.Checkout(
         id=checkout_with_existing_checkout_payment_id,
         module=TEST_MODULE_ROOT,
         name="Test Payment",
@@ -54,7 +54,7 @@ async def init_objects() -> None:
     await add_object_to_db(checkout_with_existing_checkout_payment)
 
     global existing_checkout_payment
-    existing_checkout_payment = models_payment.CheckoutPayment(
+    existing_checkout_payment = models_checkout.CheckoutPayment(
         id=uuid.uuid4(),
         checkout_id=checkout_with_existing_checkout_payment_id,
         paid_amount=100,
@@ -64,7 +64,7 @@ async def init_objects() -> None:
     await add_object_to_db(existing_checkout_payment)
 
     global checkout
-    checkout = models_payment.Checkout(
+    checkout = models_checkout.Checkout(
         id=uuid.uuid4(),
         module="tests",
         name="Test Payment",
@@ -122,7 +122,7 @@ def test_webhook_payment_for_already_received_payment(
     This situation could happen if HelloAsso call our webhook multiple times for the same payment.
     """
     mocked_hyperion_security_logger = mocker.patch(
-        "app.core.payment.endpoints_payment.hyperion_error_logger.debug",
+        "app.core.checkout.endpoints_checkout.hyperion_error_logger.debug",
     )
 
     response = client.post(
@@ -245,7 +245,7 @@ async def test_webhook_payment(
     assert response.status_code == 204
 
     async with get_TestingSessionLocal()() as db:
-        checkout_model = await cruds_payment.get_checkout_by_id(
+        checkout_model = await cruds_checkout.get_checkout_by_id(
             checkout_id=checkout.id,
             db=db,
         )
@@ -271,7 +271,7 @@ async def test_webhook_payment(
     assert response.status_code == 204
 
     async with get_TestingSessionLocal()() as db:
-        checkout_model = await cruds_payment.get_checkout_by_id(
+        checkout_model = await cruds_checkout.get_checkout_by_id(
             checkout_id=checkout.id,
             db=db,
         )
@@ -284,7 +284,7 @@ async def test_webhook_payment(
 
 
 async def callback(
-    checkout_payment: schemas_payment.CheckoutPayment,
+    checkout_payment: schemas_checkout.CheckoutPayment,
     db: AsyncSession,
 ) -> None:
     pass
@@ -296,7 +296,7 @@ async def test_webhook_payment_callback(
 ) -> None:
     # We patch the callback to be able to check if it was called
     mocked_callback = mocker.patch(
-        "tests.core.test_payment.callback",
+        "tests.core.test_checkout.callback",
     )
 
     # We patch the module_list to inject our custom test module
@@ -309,7 +309,7 @@ async def test_webhook_payment_callback(
         permissions=None,
     )
     mocker.patch(
-        "app.core.payment.endpoints_payment.all_modules",
+        "app.core.checkout.endpoints_checkout.all_modules",
         [test_module],
     )
 
@@ -338,7 +338,7 @@ async def test_webhook_payment_callback_fail(
 ) -> None:
     # We patch the callback to be able to check if it was called
     mocked_callback = mocker.patch(
-        "tests.core.test_payment.callback",
+        "tests.core.test_checkout.callback",
         side_effect=ValueError("Test error"),
     )
 
@@ -352,12 +352,12 @@ async def test_webhook_payment_callback_fail(
         permissions=None,
     )
     mocker.patch(
-        "app.core.payment.endpoints_payment.all_modules",
+        "app.core.checkout.endpoints_checkout.all_modules",
         [test_module],
     )
 
     mocked_hyperion_security_logger = mocker.patch(
-        "app.core.payment.endpoints_payment.hyperion_error_logger.exception",
+        "app.core.checkout.endpoints_checkout.hyperion_error_logger.exception",
     )
 
     response = client.post(
@@ -442,7 +442,7 @@ async def test_payment_tool_init_checkout(
         redirect_url=redirect_url,
     )
     mocker.patch(
-        "app.core.payment.payment_tool.CheckoutApi",
+        "app.core.checkout.payment_tool.CheckoutApi",
         return_value=mock_checkout_api,
     )
 
@@ -515,7 +515,7 @@ async def test_payment_tool_init_checkout_with_one_failure(
     mock_checkout_api = mocker.MagicMock()
     mock_checkout_api.organizations_organization_slug_checkout_intents_post.side_effect = init_a_checkout_side_effect
     mocker.patch(
-        "app.core.payment.payment_tool.CheckoutApi",
+        "app.core.checkout.payment_tool.CheckoutApi",
         return_value=mock_checkout_api,
     )
 
@@ -544,7 +544,7 @@ async def test_payment_tool_init_checkout_fail(
     client: TestClient,
 ) -> None:
     mocked_hyperion_security_logger = mocker.patch(
-        "app.core.payment.endpoints_payment.hyperion_error_logger.error",
+        "app.core.checkout.endpoints_checkout.hyperion_error_logger.error",
     )
 
     redirect_url = "https://example.com"
@@ -580,7 +580,7 @@ async def test_payment_tool_init_checkout_fail(
     )
 
     mocker.patch(
-        "app.core.payment.payment_tool.CheckoutApi",
+        "app.core.checkout.payment_tool.CheckoutApi",
         return_value=mock_checkout_api,
     )
