+++ tests/modules/sport_competition/test_purchases.py
@@ -5,8 +5,8 @@ import pytest
 import pytest_asyncio
 from fastapi.testclient import TestClient
 
+from app.core.checkout import models_checkout
 from app.core.groups import models_groups
-from app.core.payment import models_payment
 from app.core.schools import models_schools
 from app.core.schools.schools_type import SchoolType
 from app.core.users import models_users
@@ -538,7 +538,7 @@ async def setup():
         created_at=datetime.now(UTC),
     )
     await add_object_to_db(payment)
-    base_checkout = models_payment.Checkout(
+    base_checkout = models_checkout.Checkout(
         id=uuid4(),
         module="competition",
         name="Competition Checkout",
